+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 This changelog summarizes all changes of the RAG Web Browser
 
-
 ### 2024-11-08
 
 🚀 Features
 - Add functionality to extract content from a specific URL
 - Update README.md to include new functionality and provide examples
+
+### 2024-10-17
+
+🚀 Features
+- Handle errors when request is added to Playwright queue.
+  This will prevent the Cheerio crawler from repeating the same request multiple times.
+- Silence error: Could not parse CSS stylesheet as there is no way to fix it at our end
+- Set logLevel to INFO (debug level can be set using the `debugMode=true` input)
 
 ### 2024-10-11
 
@@ -24,7 +31,6 @@ This changelog summarizes all changes of the RAG Web Browser
 - Updated README.md to include tips on improving latency
 - Set initialConcurrency to 5
 - Set minConcurrency to 3
-- Set logLevel to INFO
 
 ### 2024-09-20